requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
python-slugify==5.0.2
six==1.16.0
SQLAlchemy==1.4.20
text-unidecode==1.3
Werkzeug==2.0.1
